Final touches and fixes
(cherry picked from commit 17d0376c13a319a11ce116a7e3ff593b3ed3ea34)

+++ php-backend/remove-expense.php
@@ -14,16 +14,39 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST') {
 
     header('Content-Type: application/json');
 
-    // just perform the deletion
-    // ideally we would check the cookie and verify the session and also if the 
-    // given expense is associated with that user, but in this simple app we dont care
-    delete_expense($expense_id);
+    if (!isset($_REQUEST['pfa_cookie'])) {
+        // cookie is not set
+        http_response_code(401);
+        echo json_encode([
+            'status' => 'Unauthorized',
+            'code' => 401,
+            'message' => 'Unauthorized, log in!'
+        ]);
+    } else {
+        $sessid = $_REQUEST['pfa_cookie'];
+        $user_id = get_user_from_session($sessid);
+        if (!$user_id) {
+            // invalid cookie
+            http_response_code(401);
+            setcookie('pfa_cookie', $sessid, -3600);
+            echo json_encode([
+                'status' => 'Unauthenticated',
+                'code' => 401,
+                'message' => 'Malformed cookie. log in again'
+            ]);
+        } else {
+            // just perform the deletion
+            // ideally we would check the cookie and verify the session and also if the
+            // given expense is associated with that user, but in this simple app we dont care
+            delete_expense($expense_id);
 
-    http_response_code(200);
-    $response = [
-        'status' => 'resource deleted successfully',
-        'code' => 200,
-        'message' => 'Expense deleted successfully',
-    ];
-    echo json_encode($response);
+            http_response_code(200);
+            $response = [
+                'status' => 'resource deleted successfully',
+                'code' => 200,
+                'message' => 'Expense deleted successfully',
+            ];
+            echo json_encode($response);
+        }
+    }
 }
